In fixed point iteration, compute min function on entire relation
Since over 20 years of research have yield so many different
semantics of min aggregation in datalog, it is worthwhile to
consider examples to help illustrate what we have in mind.

Example:

  (define-sort router_t () (_ BitVec 1))
  (define-sort cost_t () (_ BitVec 4))
  (declare-rel length (router_t cost_t))

  (declare-var node router_t)
  (declare-var cost cost_t)
  (declare-var new_cost cost_t)

  (rule (length #b0 #x2))
  (rule (length #b0 #x3))
  (rule (length #b1 #x5))

  (rule (=>
            (and (length node cost) ((_ min length 1) node cost)
                 (= (bvadd cost #x1) new_cost))
            (length node new_cost)))

  (query (length node cost) :print-answer true)

We expect the following:

(or (and (= (:var 0) #b0) (= (:var 1) #x2))
    (and (= (:var 0) #b0) (= (:var 1) #x3))
    (and (= (:var 0) #b1) (= (:var 1) #x5))
    (and (= (:var 0) #b1) (= (:var 1) #x6)))

The purpose of this example is to illustrate that the min aggregation
function is evaluated on the entire relation as opposed to the delta
used in the while-loop condition of the fixed point iteration.