Simplify SMT2 min interface and handle min over joins
Previously, the min aggregation function took a non-negative
integer to designate the column to be minimized. To simplify
the SMT2 interface, min takes now a variable name instead.
Furthermore, the user does not need to specify a relation
any longer. For this to work, we follow the LogicBlox 4.2
convention and allow only a single aggregation function in
the body of a rule where the group-by clause is inferred
from the head of the rule.

One other benefit of this design is that we can handle min
and joins in the same rule (which is needed for our benchmarks).

One downside is that any joins need to generally happen before
the min aggregation function is applied. This may incur performance
penalties when the join result is large unless we can bypass the
monotonicity requirement of relations which are subject to min.

Example:

      (define-sort nibble_t () (_ BitVec 4))
      (declare-rel triples (nibble_t nibble_t nibble_t))
      (declare-rel pairs (nibble_t nibble_t))

      (declare-var x nibble_t)
      (declare-var y nibble_t)
      (declare-var z nibble_t)

      (rule (triples #x0 #x3 #x5))
      (rule (triples #x0 #x1 #x4))

      (rule (=> (and (triples x y z) (min y)) (pairs x y)))

This says that we want to find the minimum of y in the pairs
relation grouped by the first column.

Note that this example shows that it would make to consider
moving min from the body into the head if we are so inclined
to do. This would also further emphasize that joins are meant
to happen before min is applied.

Signed-off-by: Alex Horn <t-alexh@microsoft.com>