Add conservative min aggregation checker
This checker traverses the directed graph which represents the dependencies
between uninterpreted predicates. The purpose of this graph traversal is to
check whether the instruction generated by the compiler::compile_final_min
instruction can be run when the interpreter goes from a lower stratum to
the next higher stratum. To check this, we identify strongly connected
components (SCCs) and look for edges that leave each SCC.

Our checker allows shortest paths datalog programs such as the following:

  (rule (=>
     (link x y c1)
     (path x y c1))
  )

  (rule (=>
    (and (sh_path x y c1) (link y z c2) (= (bvadd c1 c2) c3))
    (path x z c3))
  )

  (rule (=>
    (and (path x y c1) (min c1))
    (sh_path x y c1))
  )

We will likely have to relax the checker to accommodate more complex
examples.

As appropriate, some of the deleted tests will be moved to the z3test
repository (https://github.com/Z3Prover/z3test).

Signed-off-by: Alex Horn <t-alexh@microsoft.com>