Fix error under Python 3.5 ``` UnicodeDecodeError: 'ascii' codec can't decode byte 0xef in position 0: ordinal not in range(128) ```
It seems to occur when running in the OpenSUSE build environment but
it doesn't when I run outside. The fix here is to read the file as
binary as explicitly decode and encode as utf8. In Python3 we could
pass an encoding parameter to the call to ``open()`` but that option
isn't available for Python 2.

Whilst I'm here also switch to using ``with`` statements for the calls
to ``open()``. The previous code with explicit ``close()`` calls is very
bad style and also unsafe if an exception gets thrown as the files might
not get closed.