adding flag to build Z3 without direct loading of the Z3-library with Java-Bindings.
This allows the user to choose other loading mechanisms, for example with a different path for the Z3-library.
This flag is needed, if no static linking is used for the binary and also no LD_LIBRARY_PATH or other searching mechanisms.
In combination with our project, we know that libz3.so and libz3java.so are in the same directory.
The flag JAVA_EXTERNAL_LIBRARY_LOADING_FLAG allows to load libz3.so directly from the same directory than other librariees that access it.