Implement a CMake build system. This build system isn't entirely independent from the old build system yet because it doesn't know how to do
* exec_pyg_scripts()
* mk_pat_db()
* mk_all_install_tactic_cpps()
* mk_all_mem_initializer_cpps()
* mk_all_gparams_register_modules()

Because of this in order for this build system to work you need to
manually invoke ``python scripts/mk_make.py --build /some/dummy/path``
first so that the generated source files get created. The fact that the
old build system pollutes the source tree with generated files is bad
design but I've abused that here just so I can get the CMake build
system up and running. I will fix this in the future but I haven't done
it yet because it requires I that refactor the above python functions so

* they can be called from CMake
* they emit files into the build tree and not the source tree.

Currently the build succeeds on Linux using CMake 3.4.3 using
the "Unix Makefiles" generator. It is able to install files (``make
install``) but not uninstall (I will implement this later). The build
also builds the C API example.

In principle the build should work on other platforms too. I haven't
paid proper attention to the compiler flags needed for MSVC so future
work will need to set the flags correctly.