[Z3py] Fix conversion of float -0.0
Issue #464 was partially fixed by c171170 but there was still a problem
when creating an FPVal from the float -0.0 (from string '-0.0' worked
fine). Reason for the problem was the use of
float.as_integer_ratio() which does not return an accurate
representation of -0.0 because the denominator is always positive. This
fix explicitly handles -0.0.