In ``small_object_allocator`` remove early return for Debug builds on non-windows platforms.
The existing code meant that the ``SASSERT()`` calls were never checked
on non-windows platforms because they were only "syntactically"
reachable in Release builds (in which they are no-ops). The comments
suggest that this early return was for Valgrind but this doesn't seem
like a good idea. If false positives occur when running Z3 under
valgrind then a suppression file should be used rather than trying to
change Z3's behaviour. This is related to #477 .