[CMake] Modify `contrib/cmake/bootstrap.py` to do nothing except print a warning.
Now that the CMake files have been moved into their intended location
it is no longer necessary for this script to exist.

However we do not want to break out-of-tree scripts that build Z3 using
CMake to suddenly break. So the script has been modified to do nothing
except print a warning.

Eventually we should remove this script.