Fix how add_git_dir_dependency() looks for .git.
add_git_dir_dependency() adds dependencies on files in the .git
directory.  But in certain scenarios, .git is a file instead of a
directory.  Manual cmake code was written to handle this case for git
worktrees.  Unfortunately, the solution is slightly different for git
subrepositories.  Switching to using `git rev-parse` to determine actual
directory locations fixes this problem.

Fixes #1551