Align CMake option for python executable with documentation
The documentation (`README-CMake.md`) mentions `Z3_PYTHON_EXECUTABLE` as
the option which should be used to specify the python executable to be
used by Z3. Unfortunately, that variable is currently not being used in
any of the CMakeLists files and therefore ignored.

It seems to me that this is a regression introduced in
53a01a07bdf245353bd0288fc50efb2739d3d3be which updated the
`README-CMake.md` but not the CMake files in regards to this variable.