#5140
@veanes
mk_bool_app_helper has a bug:
When it simplifies a disjunction or conjunction of regex membership constraints of the form (and (str.in_re "" R) (str.in_re x Q))
then the first term (str.in_re "" R) is omitted in the result.
You have a test here
https://github.com/Z3Prover/z3/blob/3da9d91866fa18aee1281a1ea6d2d7ab55589740/src/ast/rewriter/seq_rewriter.cpp#L438
that means a regex membership with empty first argument is not put in the two buffers with membership/non-membership.
It isn't put into new_args either because the test bypasses these
https://github.com/Z3Prover/z3/blob/3da9d91866fa18aee1281a1ea6d2d7ab55589740/src/ast/rewriter/seq_rewriter.cpp#L485