use non-adjacent form for multiplying by a constant
Non-adjacent form (NAF) should never be harmful, and is more general
than checking for multiplication by -1.

In cases where the transformation doesn't help then I aim to generate
the same circuit that long multiplication would have, which I arrange by
delegating most of the work to the general long-multiplication
implementation. If there are bit sequences like 00110 within the
constant operand, then what would have been addition of two positive
terms becomes addition of one positive and one negative term--so in that
case the result isn't better and isn't identical to long multiplication,
but I think it should be close enough.

I think this could be microoptimized by checking which input bits are
constant once, and deciding between pure-constant or const-case or NAF
accordingly. But that's probably not important.